When an attempt is made to apply a hunk to a missing file, log warning.
Instead of failing the export process.  This is probably not the 'correct'
thing to do, but until I can understand how there can be two patches in the
inventory, first one removes the file, second wants to apply a hunk.

The rmfile and hunk file are NOT in merger segments (which we ignore).
There is some file manipulation logic in Darcs which I don't understand.  When
I do understand it, I will revisit this decision.